Fixed minor collision bug
When going fast up a ramp, if you tried to jump, it was possible it
could detect the next tile as higher ground (and a ceiling) and suck you
to it, resulting in you losing all momentum and not jumping. This has
been fixed by moving an if statement to a more sensical place.